🔖 Release: version 1.1.4

+++ build_deb_docker.sh
@@ -1,21 +1,23 @@
 #!/bin/bash
-# Script pour construire un paquet Debian dans Docker depuis macOS
 set -e
 
 # Vérifie que Docker est installé
-if ! command -v docker &> /dev/null; then
-  echo "Docker n'est pas installé. Installez-le d'abord." >&2
-  exit 1
-fi
+command -v docker &>/dev/null || { echo "Docker n'est pas installé. Installez-le d'abord." >&2; exit 1; }
 
-# Lance le build dans un conteneur Debian
 DOCKER_IMAGE=debian:bookworm
 PROJECT_DIR=$(pwd)
 
-# Commande à exécuter dans le conteneur
-DOCKER_CMD='apt-get update && apt-get install -y dpkg-dev && dpkg-deb --build mypackage && chown $(id -u):$(id -g) mypackage.deb'
+VERSION=$(awk '/\[workspace.package\]/ {found=1} found && /^version/ {gsub(/\"/, "", $3); print $3; exit}' Cargo.toml)
+PKG_NAME="TEV-${VERSION}.deb"
+
+DOCKER_CMD="
+  apt-get update &&
+  apt-get install -y dpkg-dev &&
+  dpkg-deb --build mypackage &&
+  mv mypackage.deb $PKG_NAME &&
+  chown $(id -u):$(id -g) $PKG_NAME
+"
 
 echo "Lancement du build dans Docker..."
 docker run --rm -v "$PROJECT_DIR":/workspace -w /workspace $DOCKER_IMAGE bash -c "$DOCKER_CMD"
-
-echo "\nFichier .deb généré dans : $PROJECT_DIR/mypackage.deb"
+echo "Fichier .deb généré dans : $PROJECT_DIR/$PKG_NAME"
